[vpr][stats] move static funcitons in stats.cpp to stats_utils
